Allow Wrio to early cancel any NA
if the paramater early_cancel is used with a Wrio attack:
Wrio will not deal any damage with that attack
Wrio will be in the normal animation state for at least early_cancel frames, and up to the full duration of the normal if not followed by any actions.
Wrio will be able to dash or jump immediately after early_cancel frames, without needing to wait on hitmark.
The normal counter will still continue.